Manual PHP future-proof fix
PHP only allows file size uploads of 2mb or less.  Since rConfig is
approaching 1.8mb in size as a .zip file, this has some manual commands
for the user to run to execute a script in the configs folder that
increases the PHP upload size to 50mb, and increases the timeout to 300
seconds.